Extract Store key-building functions into keys.rs.
Continuing to break up the store into more granular modules. We're
almost there.
