+++ src/store/cron/ops.rs
@@ -14,8 +14,9 @@ use fjall::{Readable, Slice};
 use tokio::task;
 
 use super::super::enqueue::{apply_enqueue, finalize_enqueue, prepare_enqueue};
+use super::super::keys::RecordKind;
 use super::super::options::{CronEntryOptions, ReplaceCronGroupOptions};
-use super::super::store::{RecordKind, Store, StoreEvent};
+use super::super::store::{Store, StoreEvent};
 use super::super::types::StoreError;
 use super::{CronEntry, CronGroup};
