+++ src/store/patch.rs
@@ -15,11 +15,12 @@ use fjall::Slice;
 use tokio::sync::broadcast;
 use tokio::task;
 
+use super::keys::{make_job_key, make_queue_key, make_status_key};
 use super::options::{BulkPatchOptions, PatchJobOptions};
 use super::ready_index::ReadyIndex;
 use super::scan::{JobStream, PayloadFilteredIter, build_id_stream};
 use super::scheduled_index::ScheduledIndex;
-use super::store::{Keyspaces, Store, StoreEvent, make_job_key, make_queue_key, make_status_key};
+use super::store::{Keyspaces, Store, StoreEvent};
 use super::types::{Job, JobStatus, ScanDirection, StoreError};
 
 impl PatchJobOptions {
