+++ src/store/take.rs
@@ -18,9 +18,8 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use fjall::Slice;
 use tokio::task;
 
-use super::store::{
-    Store, StoreEvent, make_job_key, make_payload_key, make_status_key, make_unique_key,
-};
+use super::keys::{make_job_key, make_payload_key, make_status_key, make_unique_key};
+use super::store::{Store, StoreEvent};
 use super::types::{Job, JobStatus, StoreError, UniqueWhile};
 
 impl Store {
